add BUILD_NUMBER (generated by Hudson) to .war filename
git-svn-id: https://mifos.svn.sourceforge.net/svnroot/mifos/trunk@14351 a8845c50-7012-0410-95d3-8e1449b9b1e4

+++ resources/rename_latest_war.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
 # renames WAR file so the revision number is part of the filename
+#
+# BUILD_NUMBER and WAR_ARCHIVE are assumed set by caller (Hudson)
 
 tdir=`mktemp -d`
 vinfo=WEB-INF/classes/org/mifos/config/resources/versionInfo.properties
@@ -16,11 +18,15 @@ cd $tdir
 jar -xf $src_war $vinfo
 # \+ requires GNU sed
 revnum=`grep revision $vinfo | sed -e 's/^[^=]*=\([[:digit:]]\+\)$/\1/'`
-dest_war=$WAR_ARCHIVE/mifos-$revnum.war
+dest_war=$WAR_ARCHIVE/mifos-$BUILD_NUMBER-$revnum.war
 
-[ -e $dest_war ] && echo $dest_war already exists and will be overwritten
+if [ -e $dest_war ]
+then
+    echo ERROR: $dest_war already exists! This should not happen.
+    exit 1
+fi
 
-mv $src_war $WAR_ARCHIVE/mifos-$revnum.war
+mv $src_war $dest_war
 
 cd /tmp
 rm -rf $tdir
